Outputs FASTQ instead of FASTA files
A quality score of 42, the highest possible quality,
with an ASCII base of 33, is now output for each read.

All command-line options that are relevant to
the Fastq output mode have now been added to
chipulate.py. This includes --fragment-length, --read-length,
--fragment-jitter options.

The default summit assignment is now changed from earlier
commits, where each region's summit is set to the middle of the
input regions.